chore(2.0): misc frontend changes
Miscellaneous frontend changes
