+++ js/package.json
@@ -8,6 +8,7 @@
     "@zxcvbn-ts/core": "^1.2.0",
     "@zxcvbn-ts/language-common": "^1.0.2",
     "@zxcvbn-ts/language-en": "^1.2.0",
+    "flarum": "^0.1.0-beta.16",
     "flarum-webpack-config": "^3.0.0",
     "webpack": "^5.65.0",
     "webpack-cli": "^4.9.1"
